--- docs/README.md
@@ -1,35 +1,0 @@
-1️⃣Car 자동차
-
-- 매개변수로 이름을 입력 받는 car 생성자 생성
-- 자동차 이름 출력 기능
-- 이름은 5글자 이하만 가능
-- 전진 기능 (위치 +1 )
-
-2️⃣Race 레이스
-
-- 매개변수로 이동 수를 입력 받는 race 생성자 생성
-- 무작위 값 생성해서 4 이상일 경우 전진
-- 무작위 값 생성해서 4 이하일 경우 멈춤(전진x)
-- 우승자 고르기
-
-3️⃣view 입출력 양식
-
-- Input 에 관한 것
-    - 문구 -> "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"
-    - 문구 -> "시도할 회수는 몇 회인가요?"
-- Output 에 관한 것
-    - '이름 : 대시' 형식
-
-4️⃣RaceController 레이스 게임 과정
-
-1. 자동차 이름 입력 받음.
-    1. 에러 경우 에러 메세지 출력 (이름이 5글자 이상일 때)
-2. 시도할 횟수를 입력 받음.
-    1. 에러 경우 에러 메세지 출력 (입력 값이 숫자가 아닐 때, 1보다 작을 때)
-3. 실행결과 알려줌
-4. 우승자 알려줌
-    1. 한 명일 경우
-    2. 여러명일 경우 -> 쉼표로 구분함
-
-
-
