+++ src/main/java/racingcar/Application.java
@@ -1,10 +1,7 @@
 package racingcar;
 
-import racingcar.controller.RacingController;
-
 public class Application {
     public static void main(String[] args) {
         // TODO 구현 진행
-        RacingController.main(args);
     }
 }
