--- src/main/java/racingcar/domain/Car.java
@@ -1,24 +1,0 @@
-package racingcar.domain;
-
-public class Car {
-    private final String name;
-    private int position = 0;
-
-    public Car(String name) {
-        this.name = name;
-    }
-
-    // 추가 기능 구현
-    public String getName() {
-        return name;
-    }
-
-    public int getPosition() {
-        return position;
-    }
-
-    public void move() {
-        this.position++;
-    }
-
-}
