--- src/main/java/racingcar/view/InputView.java
@@ -1,19 +1,0 @@
-package racingcar.view;
-
-import camp.nextstep.edu.missionutils.Console;
-import java.util.Arrays;
-import java.util.List;
-
-public class InputView {
-
-    public static List<String> getCarNames() {
-        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,) 기준으로 구분).");
-        return Arrays.asList(Console.readLine().split(","));
-    }
-
-    public static int getRaceNum() {
-        System.out.println("시도할 회수는 몇 회인가요?");
-        return Integer.parseInt(Console.readLine());
-    }
-
-}
